requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0
albumentations>=2.0.5

# Data processing
numpy>=1.20.0
PyYAML>=6.0
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0

# Development
tqdm>=4.65.0 